feat(embed): allow to use Bow in embedded schemas (#22)
Changed the `embed_as` property from :self to :dump.
This way Bow structs will become strings in embeds
just as they are strings in regular schemas.

This is far from perfect, there is no mechanism
to automatically set the scope for embedded schemas
when you use `Bow.Ecto.store` and `Bow.Ecto.url`,
so you have to do it manually and use `Bow.store`
and `Bow.url` directly. It is not obvious though
how it should work in the first place.